+++ rumqttd/src/server/broker.rs
@@ -139,8 +139,15 @@ impl Broker {
     pub fn link(&self, client_id: &str) -> Result<(LinkTx, LinkRx), local::LinkError> {
         // Register this connection with the router. Router replies with ack which if ok will
         // start the link. Router can sometimes reject the connection (ex max connection limit)
-        let (link_tx, link_rx, _ack) =
-            Link::new(None, client_id, self.router_tx.clone(), true, None, false)?;
+        let (link_tx, link_rx, _ack) = Link::new(
+            None,
+            client_id,
+            self.router_tx.clone(),
+            true,
+            None,
+            false,
+            None,
+        )?;
         Ok((link_tx, link_rx))
     }